Fixed bug where the plugin was trying to parse an string as a struct.

+++ attachments/model/callbacks.cfm
@@ -315,9 +315,11 @@
 		if (IsSimpleValue(this[arguments.property]) && IsJson(this[arguments.property]))
 			this[arguments.property] = DeserializeJson(this[arguments.property]);
 
-		loc.filePath = Replace(this[arguments.property].path, "\", "/", "all");
+		if (StructKeyExists(this[arguments.property], "path")) {
+			loc.filePath = Replace(this[arguments.property].path, "\", "/", "all");
 		
-		this[arguments.attachmentConfig.property] = $deleteFileFromStorage(source=loc.filePath, argumentCollection=arguments.attachmentConfig);
+			this[arguments.attachmentConfig.property] = $deleteFileFromStorage(source=loc.filePath, argumentCollection=arguments.attachmentConfig);
+		}
 	</cfscript>
 	<cfreturn true />
 </cffunction>
